Cleaned up the dev-navigation's edits
