+++ BITS-App/App.xaml.cs
@@ -9,8 +9,6 @@ public partial class App : Application {
 	public App() {
 		InitializeComponent();
 
-        MainPage = new NavigationPage(new Pages.HomePage());
-
         client = new HttpClient();
 
 		MainPage = new AppTabbedPage();
