--- BITS-App/AppShell.xaml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<Shell
-    x:Class="BITS_App.AppShell"
-    xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-    xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    xmlns:local="clr-namespace:BITS_App"
-    xmlns:views="clr-namespace:BITS_App.Views"
-    xmlns:pages="clr-namespace:BITS_App.Pages"
-    Shell.FlyoutBehavior="Disabled">
-
-    <ShellContent
-        ContentTemplate="{DataTemplate pages:HomePage}"
-        Route="Main" />
-
-    <ShellContent
-        ContentTemplate="{DataTemplate views:PostPage}"
-        Route="PostPage" />
-</Shell>
