+++ BITS-App/AppTabbedPage.xaml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <TabbedPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-            xmlns:views="clr-namespace:BITS_App.Views"
             xmlns:pages="clr-namespace:BITS_App.Pages"
             x:Class="BITS_App.AppTabbedPage"
             Title="AppTabbedPage"
@@ -22,7 +21,7 @@
     
     <NavigationPage Title="Staff">
         <x:Arguments>
-            <views:StaffProfilePage/>
+            <pages:StaffProfilePage/>
         </x:Arguments>
     </NavigationPage>
